Added 4 modules which deal with Data independent of the ledger to the directory semantics/executable-spec/src/Data/  These can be imported using the module names Data.CannonicalMaps.hs Data.Coders Data.MemoBytes Control.SetAlgebra
