+++ shelley/chain-and-ledger/executable-spec/src/Shelley/Spec/Ledger/STS/Utxow.hs
@@ -29,9 +29,9 @@ import qualified Cardano.Ledger.Core as Core
 import qualified Cardano.Ledger.Crypto as CryptoClass
 import Cardano.Ledger.Era (Crypto, Era)
 import Cardano.Ledger.Shelley (ShelleyBased, ShelleyEra)
-import Control.Iterate.SetAlgebra (eval, (∩))
 import Control.Monad (when)
 import Control.Monad.Trans.Reader (asks)
+import Control.SetAlgebra (eval, (∩))
 import Control.State.Transition
   ( Embed,
     IRC (..),
